+++ src/Dice.jl
@@ -1,5 +1,7 @@
 module Dice
 
+include("core.jl")
+include("dice_ite.jl")
 include("dist/dist.jl")
 include("inference/inference.jl")
 include("analysis/analysis.jl")
