+++ src/core.jl
@@ -1,5 +1,23 @@
+export flip, prob_equals, AnyBool, Dist, isdeterministic
 
-export flip, prob_equals, AnyBool
+using DirectedAcyclicGraphs
+import DirectedAcyclicGraphs: children, NodeType, DAG, Inner, Leaf
+
+# TODO make DirectedAcyclicGraphs.DAG a trait so we are not thrown off by multiple inheritance here
+"A probability distribution over values of type `T`"
+abstract type Dist{T}  <: DAG end
+
+"Does the distribution have a deterministic value?"
+isdeterministic(x) =
+    isempty(tobits(x))
+
+function Base.show(io::IO, x::Dist) 
+    if isdeterministic(x)
+        print(io, "$(typeof(x))($(frombits(x, nothing)))")
+    else
+        print(io, "$(typeof(x))@$(hash(x)÷ 10000000000000)")
+    end
+end
 
 ##################################
 # types, structs, and constructors
